+++ internal/repository/postgresql/feed_repository.go
@@ -68,7 +68,7 @@ func (r *Repository) FeedCreate(f feed.Feed) error {
 
 func (r *Repository) FeedGetBySlug(feedSlug string) (feed.Feed, error) {
 	query := `
-	SELECT uuid, feed_url, title, slug, etag, last_modified
+	SELECT uuid, feed_url, title, description, slug, etag, last_modified
 	FROM feed_feeds
 	WHERE slug=$1`
 
@@ -77,7 +77,7 @@ func (r *Repository) FeedGetBySlug(feedSlug string) (feed.Feed, error) {
 
 func (r *Repository) FeedGetByURL(feedURL string) (feed.Feed, error) {
 	query := `
-	SELECT uuid, feed_url, title, slug, etag, last_modified
+	SELECT uuid, feed_url, title, description, slug, etag, last_modified
 	FROM feed_feeds
 	WHERE feed_url=$1`
 
@@ -86,7 +86,7 @@ func (r *Repository) FeedGetByURL(feedURL string) (feed.Feed, error) {
 
 func (r *Repository) FeedGetByUUID(feedUUID string) (feed.Feed, error) {
 	query := `
-	SELECT uuid, feed_url, title, slug, etag, last_modified
+	SELECT uuid, feed_url, title, description, slug, etag, last_modified
 	FROM feed_feeds
 	WHERE uuid=$1`
 
@@ -95,7 +95,7 @@ func (r *Repository) FeedGetByUUID(feedUUID string) (feed.Feed, error) {
 
 func (r *Repository) FeedGetNByLastSynchronizationTime(n uint, before time.Time) ([]feed.Feed, error) {
 	query := `
-	SELECT f.uuid, f.feed_url, f.title, f.slug, f.etag, f.last_modified
+	SELECT f.uuid, f.feed_url, f.title, f.description, f.slug, f.etag, f.last_modified
 	FROM feed_feeds f
 	INNER JOIN feed_subscriptions fs ON f.uuid = fs.feed_uuid
 	WHERE fetched_at < $1