+++ docs/src/changelog.md
@@ -11,8 +11,8 @@ project adheres to [Semantic Versioning](https://semver.org/).
 ### Added
 #### Feeds
 - For each entry in the list, display the title of the corresponding feed
-- Add full-text search for feed entries
 - Save and display feed descriptions
+- Add full-text search based on feed and entry metadata
 
 ### Changed
 ### CI